Fix header when scrolling
The return JSX from the Header component was wrapped in a new DIV, which
acts as a placeholder for the header when it has a fixed position since
its space was occupied by the remaining elements in the body. When the
user scrolls the header adopts a fixed position, but when the user is
at the top of the document, it returns to its normal state.