Create Search component and create search section
The search section is made up of a form, which contains an input
and a DIV responsible for create the effect of increasing the light
of the INPUT when the input is selected using background-image. Due to
the fact that transition does not support background-image, the effect
is accomplished using opacity.