Create moviesInWatchList prop in App component
This prop is declared in the App component to later store the
information in firestore. This prop, as well as its method, are passed
via useContext to the movieCard component, so every time the user
clicks on the add movie to watchlist, the prop store the image of the
respective movie. Likewise, when clicking on the checked button once
the movie is already in the watchlist, it is removed from the array.